+++ src/store/base/types.js
@@ -1,5 +1,8 @@
 // Base types
 
+// Successfully loaded liac logo
+export const LOAD_LILAC_LOGO_SUCCESS = 'LOAD_LILAC_LOGO_SUCCESS';
+
 // Successfully loaded header image
 export const LOAD_PLACEHOLDER_SUCCESS = 'LOAD_PLACEHOLDER_SUCCESS';